+++ shop/shop-controller/src/main/java/shop/woowasap/shop/controller/ProductAdminController.java
@@ -1,0 +1,45 @@
+package shop.woowasap.shop.controller;
+
+import java.net.URI;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import shop.woowasap.shop.app.api.ProductUseCase;
+import shop.woowasap.shop.app.api.request.RegisterProductRequest;
+import shop.woowasap.shop.app.api.request.UpdateProductRequest;
+import shop.woowasap.shop.app.exception.CannotFindProductException;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/v1/admin/products")
+public class ProductAdminController {
+
+    private final ProductUseCase productUseCase;
+
+    @PutMapping("/{product-id}")
+    public ResponseEntity<Void> updateProduct(@PathVariable("product-id") Long productId,
+        @RequestBody UpdateProductRequest request) {
+        productUseCase.update(productId, request);
+        return ResponseEntity.ok().build();
+    }
+
+    @PostMapping
+    public ResponseEntity<Void> registerProduct(
+        @RequestBody final RegisterProductRequest registerProductRequest) {
+        final Long registeredProductId = productUseCase.registerProduct(registerProductRequest);
+
+        return ResponseEntity.created(URI.create("/v1/admin/products/" + registeredProductId))
+            .build();
+    }
+
+    @ExceptionHandler(CannotFindProductException.class)
+    public ResponseEntity<Void> handleException() {
+        return ResponseEntity.badRequest().build();
+    }
+}
